Refactoring slack endpoints and introducing slack verification middleware
