+++ routes/index.js
@@ -1,0 +1,7 @@
+const express = require('express');
+const router = express.Router();
+
+const statusRoute = require('./status');
+router.use('/status', statusRoute);
+
+module.exports = router;
